moved methods into IMobileUtils
