+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/utils/mobile/IMobileUtils.java
@@ -864,4 +864,92 @@ public interface IMobileUtils extends IDriverPool {
         return drv;
     }
 
+    // TODO Update this method using findByImage strategy
+    /**
+     * Pressing bottom right button on the keyboard by coordinates: "search", "ok",
+     * "next", etc. - various keys appear at this position. Tested at Nexus 6P
+     * Android 8.0.0 standard keyboard. Coefficients of coordinates for other
+     * devices and custom keyboards could be different.
+     */
+    @SuppressWarnings("rawtypes")
+    default public void pressBottomRightKey() {
+        WebDriver driver = castDriver();
+        Dimension size = helper.performIgnoreException(() -> driver.manage().window().getSize());
+        int height = size.getHeight();
+        int width = size.getWidth();
+
+        PointOption<?> option = PointOption.point((int) (width * 0.915), (int) (height * 0.945));
+        new TouchAction((MobileDriver<?>) castDriver()).tap(option).perform();
+    }
+
+    default public boolean isChecked(final ExtendedWebElement element) {
+        // TODO: SZ migrate to FluentWaits
+        return element.isElementPresent(5)
+                && (element.getElement().isSelected() || element.getAttribute("checked").equals("true"));
+    }
+
+    /**
+     * If the application you're interested about is installed - returns "true".
+     * Otherwise, returns "false".
+     *
+     * @param packageName
+     *            - app's package or bundle id
+     * 
+     * @return boolean
+     */
+    default public boolean isApplicationInstalled(String packageName) {
+        boolean installed = ((MobileDriver<?>) castDriver()).isAppInstalled(packageName);
+        UTILS_LOGGER.info(String.format("Application by package name (%s) installed: ", packageName) + installed);
+        return installed;
+    }
+
+    /**
+     * Method to launch Android application by its package name.
+     *
+     * Application should be installed to device.
+     *
+     * Application might not be running in background, but will be launched anyway.
+     *
+     * @param packageName
+     *            - app's package or bundle id
+     */
+    default public void startApp(String packageName) {
+        UTILS_LOGGER.info("Starting " + packageName);
+        ((MobileDriver<?>) castDriver()).activateApp(packageName);
+    }
+
+    /**
+     * Will install application if path to apk-file on working machine is set.
+     *
+     * @param apkPath String
+     */
+    default public void installApp(String apkPath) {
+        UTILS_LOGGER.info("Will install application with apk-file from " + apkPath);
+        ((MobileDriver<?>) castDriver()).installApp(apkPath);
+    }
+
+    /**
+     * To remove installed application by provided package name
+     *
+     * @param packageName
+     *            - app's package or bundle id
+     *
+     * @return true if succeed
+     */
+    default public boolean removeApp(String packageName) {
+        boolean removed = ((MobileDriver<?>) castDriver()).removeApp(packageName);
+        UTILS_LOGGER.info(String.format("Application (%s) is successfuly removed: ", packageName) + removed);
+        return removed;
+    }
+
+    /**
+     * Method to reset test application.
+     *
+     * App's settings will be reset. User will be logged out. Application will be
+     * closed to background.
+     */
+    default public void clearAppCache() {
+        UTILS_LOGGER.info("Initiation application reset...");
+        ((MobileDriver<?>) castDriver()).resetApp();
+    }
 }
